chore: 폴더명 변경 및 로그아웃 버튼 구현
